refactor: move last remnants of uicode.js to main.js and metric-legend.js
