+++ README.md
@@ -2,8 +2,6 @@
 
 Graphing interface for MetricQ
 
-TODO: Refactoring Javascript Source code (especially in uicode.js)
-
 ## code style
 
 We use [standardJS](https://standardjs.com), but since we need some global variables, the linter should be called like this:
